rework container names w explicit arch
