+++ scripts/ci/docker/alpinedav_ubuntu_20.04_devel/build.sh
@@ -4,7 +4,8 @@
 # other details. No copyright assignment is required to contribute to Ascent.
 set -ev
 
-export TAG_NAME=alpinedav/ascent-ci:ubuntu-20.04-devel
+export TAG_ARCH=`uname -m`
+export TAG_NAME=alpinedav/ascent-ci:ubuntu-20.04-${ARCH}-devel
 
 # exec docker build to create image
 echo "docker build -t ${TAG_NAME} ."
