+++ src/examples/docker/ubuntu/ascent_jupyter/Dockerfile
@@ -2,7 +2,8 @@
 # Project developers. See top-level LICENSE AND COPYRIGHT files for dates and
 # other details. No copyright assignment is required to contribute to Ascent.
 
-FROM alpinedav/ascent:latest
+ARG TAG=latest
+FROM alpinedav/ascent:${TAG}
 
 # add our 'user' with password 'docker'
 RUN useradd -ms /bin/bash -G sudo user && echo "user:docker" | chpasswd
